requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
gitdb==4.0.7
GitPython==3.1.18
gunicorn==20.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
smmap==4.0.0
sqlparse==0.4.2
gunicorn
whitenoise
